added the application (Flask) with the JS and HTML interface. Reorganized folders for ETL and ML_Data.
