+++ AIS_Application/templates/index.html
@@ -1,0 +1,126 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>AIS Vessel Efforts</title>
+    <link rel="stylesheet" href="https://unpkg.com/leaflet@1.8.0/dist/leaflet.css" integrity="sha512-hoalWLoI8r4UszCkZ5kL8vayOGVae1oxXe/2A4AO6J9+580uKHDO3JdHb7NzwwzK5xr/Fs0W40kiNHxM9vyTtQ==" crossorigin="" />
+    <link rel="stylesheet" href="static/css/style.css">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
+</head>
+<body style="background-color:rgb(145, 178, 178)">
+    <div class="wrapper">
+        <nav class="navbar navbar-dark bg-dark navbar-expand-lg">
+            <a class="navbar-brand" href="index.html"><small>Refresh Search</small></a>
+        </nav>
+    </div>
+    <div style="background-color: rgb(133, 147, 149);">
+    <div id="jumbo" class="jumbotron col-lg-12">
+        <h1 style="color:whitesmoke; text-align: left; text-shadow: -2px -2px 0 black;">San Francisco AIS Vessel Efforts</h1>
+    </div>
+    </div>
+        <div class="progress" style="background-color: rgb(145, 161, 175)">
+        <div id="progress-bar" class="progress-bar" role="progressbar" aria-valuemin="0" aria-valuemax="100"></div> 
+        <!-- style="width: 25%" aria-valuenow="25" -->
+        </div>
+    <div class="container-fluid" style="background-color:rgb(125, 150, 150)">
+        <div class="row">
+            <div class="col-md-12">
+            <h3>AIS Data Breakdown<hr><small>Unique Vessels, Percentage of Overall Activity, and Trends Over Time<small><small><small><hr>**Note: Larger date-intervals require longer loading times. If running on slower connections, consider changing the search criteria.**</small></small></small></small></h3>
+            </div>
+        </div>
+    </div>
+    <div class="container-fluid" style="background-color:rgb(145, 178, 178)">
+        <div class="row" id="heading">
+            <div class="col-sm-2" style="background-color:rgb(125, 150, 150); border: 2px solid black; color: black">
+                <!-- <div>
+                <form>
+                        <h5>Search Criteria:</h5>
+                    <ul class="list-group no-bullets">
+                        <li class="list-group">
+                            <label for="date_start">Enter Start Date</label>
+                        <!-- </li>
+                        <li> -->
+                            <!-- <input type="datetime" placeholder="1/1/2021" id="date_start" />
+                        </li>
+                        <li class="list-group">
+                            <label for="date_end">Enter End Date</label> -->
+                        <!-- </li>
+                        <li> -->
+                            <!-- <input type="datetime" placeholder="12/31/2021" id="date_end" />
+                        </li>
+                    </ul>
+                </form>
+                <button id="search_input">Submit Query</button>
+                </div> -->
+                <div>
+                    <form name="latlon" id="latlon" class="list-group no-bullets">
+                        <li class="list-group">Top Bound:</li>
+                        <input type="text" class="list-group" name="topbound" id="topbound"></input>
+                        <li class="list-group">Bottom Bound:</li>
+                        <input type="text" class="list-group" name="bottombound" id="bottombound"></input>
+                        <li class="list-group">Left Bound:</li>
+                        <input type="text" class="list-group" name="leftbound" id="leftbound"></input>
+                        <li class="list-group">Right Bound:</li>
+                        <input type="text" class="list-group" name="rightbound" id="rightbound"></input>
+                        <button type="submit">submit</button>
+                    </form>
+                </div>
+            </div>
+            <div class="col-sm-8" id="map"></div>
+            <div class="col-sm-2" style="background-color: rgb(125, 150, 150); border: 2px solid black">
+                <div id="search-desc" style="color:rgb(162, 41, 41); font-size: small;"></div>
+                <div id="area-desc" style="color:blue; font-size: small;"></div>
+                <div id="status-update" style="font-size: small; color:black"></div>
+            </div>
+        </div>
+    </div>
+    
+    <script src="https://unpkg.com/leaflet@1.8.0/dist/leaflet.js" integrity="sha512-BB3hKbKWOc9Ez/TAwyWxNXeoV9c1v6FIeYiBieIWkpLjauysF18NzgR1MBNBXf8/KABdlkX68nAhlwcDFLGPCQ==" crossorigin=""></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/d3/4.11.0/d3.js"></script>
+    <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
+    <!-- <script src="test_list.json"></script> -->
+    <script src="static/JS/config.js"></script>
+    <script src="static/JS/logic.js"></script>
+    <!-- <script src="app.py"></script> -->
+       <!--Jquery Cdn -->
+      <script src="https://code.jquery.com/jquery-3.5.1.js"
+              integrity="sha256-QWo7LDvxbWT2tbbQ97B53yJnYU3WhH/C8ycbRAkjPDc="
+              crossorigin="anonymous"></script>
+
+      <script type="text/javascript">
+        $(document).on('submit','#latlon',function(e)
+                       {
+          console.log('hello');
+          $.ajax({
+            type:'POST',
+            url:'/',
+            data: JSON.stringify({
+              Top_Bound:$("#topbound").val(),
+              Bottom_Bound:$("#bottombound").val(),
+              Left_Bound:$("#leftbound").val(),
+              Right_Bound:$("#rightbound").val()}),
+            contentType: 'application/json',
+            dataType: 'json',
+            type: 'post',
+
+            success:function(data)
+            {
+              $('#response pre').html( JSON.stringify( data ) );
+              alert('saved');
+            }
+          }).done(function (data) {
+             console.log(data)});
+
+          e.preventDefault();
+
+        });
+      </script>
+
+    
+
+
+
+</body>
+</html>
