--- Front_end/Static/JS/config.js
@@ -1,2 +1,0 @@
-// API key
-const API_KEY = "pk.eyJ1IjoibmFidWNoIiwiYSI6ImNsMmd3eHgyeTAwcWczaW1oeGR5dDg5bHQifQ.fhSmOPs6kjzTh8K6ifqr2w";
