--- Front_end/Static/css/style.css
@@ -1,22 +1,0 @@
-.jumbotron {
-    background-image: url("../Images/golden_gate_pexels.jpg");
-    background-size: 100% 100%;
-    text-align: bottom;}
-
-#jumbo {
-    margin: auto;
-}
-
-#heading {
-    height: 375px;
-}
-
-.mb-3 {
-    margin-bottom: 10px;
-    margin-top: 200px;
-
-}
-
-.mb-4 {
-    margin-top: 0px;
-}
