+++ ML_Data/ml_data_monthly_2018.csv
@@ -1,0 +1,13 @@
+,Fishing,TugTow,Recreational,Passenger,Cargo,Tanker,Other,Unavailable
+2018/1,40,66,414,89,189,68,56,54
+2018/2,35,69,448,87,185,56,54,52
+2018/3,32,64,470,94,198,60,50,60
+2018/4,31,63,542,97,198,70,54,76
+2018/5,31,66,562,103,194,64,57,73
+2018/6,34,69,574,96,193,64,67,65
+2018/7,42,71,624,93,203,61,52,81
+2018/8,45,73,634,92,196,65,56,91
+2018/9,39,82,699,102,177,50,72,92
+2018/10,37,78,610,110,183,59,90,124
+2018/11,39,77,487,99,176,46,66,92
+2018/12,38,66,443,93,184,55,59,84
