+++ ML_Data/ml_data_monthly_2019.csv
@@ -1,0 +1,13 @@
+,Fishing,TugTow,Recreational,Passenger,Cargo,Tanker,Other,Unavailable
+2019/1,28,72,478,92,177,53,60,95
+2019/2,20,69,406,93,169,64,56,94
+2019/3,26,70,523,97,163,63,58,98
+2019/4,36,73,616,105,156,67,65,113
+2019/5,38,69,679,101,162,62,72,121
+2019/6,44,73,649,95,153,61,63,120
+2019/7,37,74,635,93,170,64,63,112
+2019/8,40,77,674,94,176,59,65,110
+2019/9,31,79,739,109,154,68,77,125
+2019/10,38,81,708,115,157,63,96,121
+2019/11,32,78,566,99,153,61,69,102
+2019/12,44,76,452,86,159,68,56,100
